+++ src/core/properties.ts
@@ -1,4 +1,5 @@
 export const CHILDREN_SORT_PROPERTY = "folderNodeChildrenSort";
 export const SIBLING_RANK_PROPERTY = "folderNodeSiblingRank";
+export const HIDDEN_PROPERTY = "folderNodeHidden";
 
 export const ICON_PROPERTY = "icon";